Adiciona a documentao de todos os métodos do projeto.

+++ list.c
@@ -1,7 +1,16 @@
 #include <stdio.h>
 #include "list.h"
-
-node init_node(int x, int y, float weight) {
+/**
+ * Cria um novo nó na posição (x,y) e com o pesso
+ * weight.
+ * O nó criado é inicializado com valores padões para
+ * next, parent, f, g e h.
+ * @param x - a posição x do nó.
+ * @param y - a posição y do nó.
+ * @param weight  - o peso do nó.
+ * @return  - o nó criado.
+ */
+node init_node(int x, int y, int weight) {
     node n;
     n.next = NULL;
     n.parent = NULL;
@@ -14,6 +23,11 @@ node init_node(int x, int y, float weight) {
     return n;
 }
 
+/**
+ * Insere um nó na lista.
+ * @param list - lista onde o nó será inserido.
+ * @param n - o nó que será inserido.
+ */
 void insert_node(node **list, node *n) {
     if (*list == NULL) {
         *list = n;
@@ -27,6 +41,11 @@ void insert_node(node **list, node *n) {
 
 }
 
+/**
+ * Remove um n ó da lista.
+ * @param list - lista de onde o nó será removido.
+ * @param n - o nó que será removido.
+ */
 void remove_node(node **list, node *n) {
     if (*list != NULL) {
         //lista vazia
@@ -53,7 +72,12 @@ void remove_node(node **list, node *n) {
     }
 }
 
-
+/**
+ * Verifica se dado nó esta contido na lista.
+ * @param list - a lista que será consultada.
+ * @param n  - o nó para a consulta.
+ * @return  1 se a lista contém o nó, caso contrário, 0.
+ */
 int contains_node(node **list, node *n) {
     if (*list == NULL) {
         return 0;
@@ -71,6 +95,10 @@ int contains_node(node **list, node *n) {
     return 0;
 }
 
+/**
+ * Imprime uma lista.
+ * @param list a lista a ser impressa.
+ */
 void print_list(node *list) {
     if (list == NULL) {
         printf("Empty list\n");
